feat: Add functionality to delete all links for a device
Introduces a new feature that allows users to remove all connections originating from or terminating at a specific device. This is achieved by adding a new action `deleteAllLinksForDevice` to the network manager and exposing it to the UI. A corresponding "Cut" icon is also added to the NetworkDiagram component to visually represent this new capability.